# swagger.yaml
openapi: 3.0.0
info:
  title: ZOZO API
  contact: {}
  version: '1.0'
servers:
- url: https://busy-red-zebra-robe.cyclic.app/api/v1
  variables: {}
paths:
...
  /products:
    get:
      tags:
      - Products
      summary: get all products
      operationId: getallproducts
      parameters:
      - name: title
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: imagesAmount
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 4
      - name: numericFilters
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: price>19,price<100000,numOfReviews>=1
      - name: limit
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 3
      - name: page
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
          example: 1
      - name: sort
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
          example: -price
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    post:
      tags:
      - Products
      summary: create product
      operationId: createproduct
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/createproductrequest'
              - example:
                  description: lorem ipsum description
                  title: '11111'
                  price: 200
            example:
              description: lorem ipsum description
              title: '11111'
              price: 200
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /products/uploadImage:
    post:
      tags:
      - Products
      summary: upload image
      operationId: uploadimage
      parameters: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            encoding: {}
            schema:
              required:
              - images
              type: object
              properties:
                images:
                  type: string
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /products/{productID}:
    parameters:
      - in: path
        name: productID
        schema:
          type: string
        required: true
        description: The product ID
    get:
      tags:
      - Products
      summary: get single products
      operationId: getsingleproducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    patch:
      tags:
      - Products
      summary: update product
      operationId: updateproduct
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/updateproductrequest'
              - example:
                  description: new desc
                  title: new title 2
            example:
              description: new desc
              title: new title 2
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Products
      summary: delete product
      operationId: deleteproduct
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    createproductrequest:
      title: createproductrequest
      required:
      - description
      - title
      - price
      type: object
      properties:
        description:
          type: string
        title:
          type: string
        price:
          type: integer
          format: int32
      example:
        description: lorem ipsum description
        title: '11111'
        price: 200
    updateproductrequest:
      title: updateproductrequest
      required:
      - description
      - title
      type: object
      properties:
        description:
          type: string
        title:
          type: string
      example:
        description: new desc
        title: new title 2
...
security:
- {}
...